Move node position / line length to separate file
